fix(assignments): move helper to service
